grid/search/models.py
from django.db import models
from rest_framework import serializers


# Create your models here.

class User(models.Model):
    first_name = models.CharField(max_length=100)
    last_name = models.CharField(max_length=100)
    age = models.IntegerField()
    gender = models.CharField(max_length=10)
    email = models.EmailField()
    phone = models.IntegerField()
    dob = models.DateField()

